move headers.js to middleware util folder
