+++ src/middleware/auth.js
@@ -15,7 +15,7 @@ limitations under the License.
 */
 import rejectingFetch from './utils/rejectingFetch';
 import invariant from 'invariant';
-import { headersGet, headersPost, headersPut, headersDelete } from './headers';
+import { headersGet, headersPost, headersPut, headersDelete } from './utils/headers';
 import { authPath } from './paths';
 
 const authFetch = (...args) => {
