+++ src/middleware/csv.js
@@ -15,7 +15,7 @@
  */
 import rejectingFetch from './utils/rejectingFetch';
 import invariant from 'invariant';
-import { headersGet, headersPost, headersPut, headersDelete } from './headers';
+import { headersGet, headersPost, headersPut, headersDelete } from './utils/headers';
 import { extensionApiPath } from './paths';
 import uploadFiles from './utils/uploadFiles';
