+++ src/middleware/order.js
@@ -16,7 +16,7 @@ limitations under the License.
 import rejectingFetch from './utils/rejectingFetch';
 import invariant from 'invariant';
 import { orderApiPath } from './paths';
-import { headersGet, headersPost, headersPut, headersDelete } from './headers';
+import { headersGet, headersPost, headersPut, headersDelete } from './utils/headers';
 import Order from '../models/Order';
 
 // likely want a registry like for inventory and hit their respective functions for each foundry
