+++ src/middleware/reporting.js
@@ -15,7 +15,7 @@
  */
 import rejectingFetch from './utils/rejectingFetch';
 import invariant from 'invariant';
-import { headersPost} from './headers';
+import { headersPost} from './utils/headers';
 import { reportApiPath } from './paths';
 
 export const reportError = (title, description, url, user) => {
