+++ src/middleware/sequence.js
@@ -15,7 +15,7 @@ limitations under the License.
 */
 import rejectingFetch from './utils/rejectingFetch';
 import { getItem, setItem } from './localStorageCache';
-import { headersGet, headersPost } from './headers';
+import { headersGet, headersPost } from './utils/headers';
 import { dataApiPath } from './paths';
 
 const getSequenceUrl = (md5, format = null, blockId = null, projectId) => dataApiPath(`sequence/${md5}` + (!!blockId ? `/${blockId}` : ''));
